requirements.txt
amqp==5.3.1
attrs==25.3.0
asgiref==3.11.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
psycopg2-binary==2.9.11
iniconfig==2.3.0
kombu==5.6.2
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
pillow==12.1.1
pyusb==1.2.1
pluggy==1.6.0
prompt_toolkit==3.0.52
Pygments==2.19.2
python-dateutil==2.9.0.post0
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
structlog==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
